requirements.txt
Django==1.11.7
djangorestframework==3.7.0
markdown==2.6.9
django-filter=1.0.4
django-rest-swagger==2.1.2
pygments==2.2.0
enum34==1.1.6
django-filter==1.1.0
django-enumfields==0.9.0
scikit-learn==0.19.0